Final Coding // 코딩 끝

+++ Assets/Script/BulletScript/EnemyBullet.cs
@@ -10,7 +10,7 @@ public class EnemyBullet : MonoBehaviour {
     public bool isDie = false;
 
 	//포탄 공격력 (기본공격력 3)
-	public int damage = 3;
+	public float damage = 3.0f;
 
     void OnEnable () {
         //Rb.AddForce(transform.forward * speed);
